Ji/gradio_run/requirements.txt
accelerate==1.1.1
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
audioop-lts==0.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Deprecated==1.2.14
diffusers==0.31.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.5.0
gradio_client==1.4.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matrix-client==0.4.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.7.3
safehttpx==0.1.1
safetensors==0.4.5
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.32.0
websockets==12.0
wheel==0.44.0
wrapt==1.16.0
zipp==3.21.0
